+++ Cargo.toml
@@ -58,10 +58,10 @@ path = "src/10_12/part_2.rs"
 
 [[bin]]
 name = "12_12_p1"
-path = "src/12_12/part_1.rs"
+path = "src/12_12/p1/part_1.rs"
 
 [[bin]]
 name = "12_12_p2"
-path = "src/12_12/part_2.rs"
+path = "src/12_12/p2/part_2.rs"
 
 [dependencies]
