+++ src/12_12/p1/way/point.rs
