+++ src/12_12/p1/way/tree.rs
