+++ src/12_12/p2/way/mod.rs
@@ -1,0 +1,4 @@
+pub mod tree;
+pub mod point;
+
+
